Keep sensitivity studies on analytical adaptation laws
Preserve the OU-III OFAT and coupled robustness study, but make the coupled sigma and tau directions follow the deployed SpectralMSE law. The sigma-coupled arm now scales r_S with exponent 6/7; the tau-coupled arm uses the exact tau^(24/7)/sqrt(T_S) ratio, reducing to tau^(41/14) away from cadence clamps. Historical robustness code is retained separately so the committed legacy evidence can still be restated byte-for-byte.

Restore a PhysicalMSE-only OU-II coefficient scale sweep with the analytical C_P as its reference, and restore the generic tuner/EMA/variance scheduling checks that were accidentally removed with the empirical-law branches.
